Launcher refactor
