+++ lib/PuppeteerSharp.Tests/ChromiumSpecificLauncherTests/BrowserUrlOptionTests.cs
@@ -1,0 +1,68 @@
+﻿using System.Threading.Tasks;
+using Xunit;
+using Xunit.Abstractions;
+
+namespace PuppeteerSharp.Tests.ConnectionsTests
+{
+    [Collection("PuppeteerLoaderFixture collection")]
+    public class BrowserUrlOptionTests : PuppeteerPageBaseTest
+    {
+        public BrowserUrlOptionTests(ITestOutputHelper output) : base(output)
+        {
+        }
+
+        [Fact]
+        public async Task ShouldBeAbleToConnectUsingBrowserURLWithAndWithoutTrailingSlash()
+        {
+            var options = TestConstants.DefaultBrowserOptions();
+            options.Args = new string[] { "--remote-debugging-port=21222" };
+            var originalBrowser = await Puppeteer.LaunchAsync(options);
+            var browserURL = "http://127.0.0.1:21222";
+
+            var browser1 = await Puppeteer.ConnectAsync(new ConnectOptions { BrowserURL = browserURL });
+            var page1 = await browser1.NewPageAsync();
+            Assert.Equal(56, await page1.EvaluateExpressionAsync<int>("7 * 8"));
+            browser1.Disconnect();
+
+            var browser2 = await Puppeteer.ConnectAsync(new ConnectOptions { BrowserURL = browserURL + "/" });
+            var page2 = await browser2.NewPageAsync();
+            Assert.Equal(56, await page2.EvaluateExpressionAsync<int>("7 * 8"));
+            browser2.Disconnect();
+            await originalBrowser.CloseAsync();
+        }
+
+        [Fact]
+        public async Task ShouldThrowWhenUsingBothBrowserWSEndpointAndBrowserURL()
+        {
+            var options = TestConstants.DefaultBrowserOptions();
+            options.Args = new string[] { "--remote-debugging-port=21222" };
+            var originalBrowser = await Puppeteer.LaunchAsync(options);
+            var browserURL = "http://127.0.0.1:21222";
+
+            await Assert.ThrowsAsync<PuppeteerException>(() => Puppeteer.ConnectAsync(new ConnectOptions
+            {
+                BrowserURL = browserURL,
+                BrowserWSEndpoint = originalBrowser.WebSocketEndpoint
+            }));
+
+            await originalBrowser.CloseAsync();
+        }
+
+        [Fact]
+        public async Task ShouldThrowWhenTryingToConnectToNonExistingBrowser()
+        {
+            var options = TestConstants.DefaultBrowserOptions();
+            options.Args = new string[] { "--remote-debugging-port=21222" };
+            var originalBrowser = await Puppeteer.LaunchAsync(options);
+            var browserURL = "http://127.0.0.1:2122";
+
+            await Assert.ThrowsAsync<ChromiumProcessException>(() => Puppeteer.ConnectAsync(new ConnectOptions
+            {
+                BrowserURL = browserURL
+            }));
+
+            await originalBrowser.CloseAsync();
+        }
+
+    }
+}
