+++ lib/PuppeteerSharp.Tests/ConnectionTests/ConnectionTests.cs
@@ -34,59 +34,6 @@ namespace PuppeteerSharp.Tests.ConnectionsTests
             Assert.False(Page.Client.HasPendingCallbacks());
         }
 
-        [Fact]
-        public async Task ShouldBeAbleToConnectUsingBrowserURLWithAndWithoutTrailingSlash()
-        {
-            var options = TestConstants.DefaultBrowserOptions();
-            options.Args = new string[] { "--remote-debugging-port=21222" };
-            var originalBrowser = await Puppeteer.LaunchAsync(options);
-            var browserURL = "http://127.0.0.1:21222";
-
-            var browser1 = await Puppeteer.ConnectAsync(new ConnectOptions { BrowserURL = browserURL });
-            var page1 = await browser1.NewPageAsync();
-            Assert.Equal(56, await page1.EvaluateExpressionAsync<int>("7 * 8"));
-            browser1.Disconnect();
-
-            var browser2 = await Puppeteer.ConnectAsync(new ConnectOptions { BrowserURL = browserURL + "/" });
-            var page2 = await browser2.NewPageAsync();
-            Assert.Equal(56, await page2.EvaluateExpressionAsync<int>("7 * 8"));
-            browser2.Disconnect();
-            await originalBrowser.CloseAsync();
-        }
-
-        [Fact]
-        public async Task ShouldThrowWhenUsingBothBrowserWSEndpointAndBrowserURL()
-        {
-            var options = TestConstants.DefaultBrowserOptions();
-            options.Args = new string[] { "--remote-debugging-port=21222" };
-            var originalBrowser = await Puppeteer.LaunchAsync(options);
-            var browserURL = "http://127.0.0.1:21222";
-
-            await Assert.ThrowsAsync<PuppeteerException>(() => Puppeteer.ConnectAsync(new ConnectOptions
-            {
-                BrowserURL = browserURL,
-                BrowserWSEndpoint = originalBrowser.WebSocketEndpoint
-            }));
-
-            await originalBrowser.CloseAsync();
-        }
-
-        [Fact]
-        public async Task ShouldThrowWhenTryingToConnectToNonExistingBrowser()
-        {
-            var options = TestConstants.DefaultBrowserOptions();
-            options.Args = new string[] { "--remote-debugging-port=21222" };
-            var originalBrowser = await Puppeteer.LaunchAsync(options);
-            var browserURL = "http://127.0.0.1:2122";
-
-            await Assert.ThrowsAsync<ChromiumProcessException>(() => Puppeteer.ConnectAsync(new ConnectOptions
-            {
-                BrowserURL = browserURL
-            }));
-
-            await originalBrowser.CloseAsync();
-        }
-
         private async Task TheSourceOfTheProblems() => await Page.Client.SendAsync("ThisCommand.DoesNotExist");
     }
 }
